# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"GetSeasonTrainingInfoResponseModel": {
				"description": "Get season training info response model for tsoa documentation",
				"properties": {
					"id": {
						"type": "string"
					},
					"season_number": {
						"type": "number",
						"format": "double"
					},
					"team_id": {
						"type": "string"
					},
					"training_facility_level": {
						"type": "number",
						"format": "double"
					},
					"training_credits_used": {
						"type": "number",
						"format": "double"
					},
					"training_credits_available": {
						"type": "number",
						"format": "double"
					},
					"training_credits_remaining": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Season training info retrieved successfully"
					}
				},
				"required": [
					"id",
					"season_number",
					"team_id",
					"training_facility_level",
					"training_credits_used",
					"training_credits_available",
					"training_credits_remaining",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SeasonTrainingInfo": {
				"properties": {
					"id": {
						"type": "string"
					},
					"season_number": {
						"type": "number",
						"format": "double"
					},
					"team_id": {
						"type": "string"
					},
					"training_facility_level": {
						"type": "number",
						"format": "double"
					},
					"training_credits_used": {
						"type": "number",
						"format": "double"
					},
					"training_credits_available": {
						"type": "number",
						"format": "double"
					},
					"training_credits_remaining": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"season_number",
					"team_id",
					"training_facility_level",
					"training_credits_used",
					"training_credits_available",
					"training_credits_remaining"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrainPlayerResponseModel": {
				"description": "Train player response model",
				"properties": {
					"player": {
						"$ref": "#/components/schemas/Player",
						"description": "Updated player data"
					},
					"season": {
						"$ref": "#/components/schemas/SeasonTrainingInfo",
						"description": "Season training information"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Player trained successfully"
					}
				},
				"required": [
					"player",
					"season",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TrainPlayerRequestModel": {
				"description": "Train player request model",
				"properties": {
					"player_id": {
						"type": "string",
						"description": "Player ID to train",
						"example": "550e8400-e29b-41d4-a716-446655440000"
					},
					"stat_name": {
						"$ref": "#/components/schemas/PlayerStatName",
						"description": "Stat name to train",
						"example": "throwing"
					}
				},
				"required": [
					"player_id",
					"stat_name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetSeasonScoutingInfoResponseModel": {
				"description": "Get season scouting info response model for tsoa documentation",
				"properties": {
					"id": {
						"type": "string"
					},
					"season_number": {
						"type": "number",
						"format": "double"
					},
					"team_id": {
						"type": "string"
					},
					"scout_level": {
						"type": "number",
						"format": "double"
					},
					"scouting_credits_used": {
						"type": "number",
						"format": "double"
					},
					"scouting_credits_available": {
						"type": "number",
						"format": "double"
					},
					"scouting_credits_remaining": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Season scouting info retrieved successfully"
					}
				},
				"required": [
					"id",
					"season_number",
					"team_id",
					"scout_level",
					"scouting_credits_used",
					"scouting_credits_available",
					"scouting_credits_remaining",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetScoutedPlayersResponseModel": {
				"description": "Get scouted players response model",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Scouted players retrieved successfully"
					},
					"scouted_players": {
						"items": {
							"$ref": "#/components/schemas/ScoutedPlayer"
						},
						"type": "array",
						"description": "List of scouted players"
					}
				},
				"required": [
					"success",
					"message",
					"scouted_players"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ScoutPlayersResponseModel": {
				"description": "Scout players response model",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Scouted players generated successfully"
					},
					"scouted_players": {
						"items": {
							"$ref": "#/components/schemas/ScoutedPlayer"
						},
						"type": "array",
						"description": "List of newly scouted players"
					}
				},
				"required": [
					"success",
					"message",
					"scouted_players"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ScoutPlayersRequestModel": {
				"description": "Scout players request model",
				"properties": {
					"count": {
						"type": "number",
						"format": "double",
						"description": "Number of players to scout",
						"example": 3
					}
				},
				"required": [
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PurchaseScoutedPlayerResponseModel": {
				"description": "Purchase scouted player response model",
				"properties": {
					"result": {
						"properties": {
							"team_budget": {
								"type": "number",
								"format": "double",
								"description": "Updated team budget",
								"example": 500
							},
							"player": {
								"$ref": "#/components/schemas/Player",
								"description": "Purchased player details"
							}
						},
						"required": [
							"team_budget",
							"player"
						],
						"type": "object",
						"description": "Result of the purchase operation"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Scouted player purchased successfully"
					}
				},
				"required": [
					"result",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PurchaseScoutedPlayerRequestModel": {
				"description": "Purchase scouted player request model",
				"properties": {
					"scouted_player_id": {
						"type": "string",
						"description": "ID of the scouted player to purchase",
						"example": "550e8400-e29b-41d4-a716-446655440000"
					}
				},
				"required": [
					"scouted_player_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CurrentGameResponseModel": {
				"description": "Current game response model for tsoa documentation",
				"properties": {
					"game_id": {
						"type": "string",
						"description": "Game ID",
						"example": "550e8400-e29b-41d4-a716-446655440000"
					},
					"game_season": {
						"type": "number",
						"format": "double",
						"description": "Game season number",
						"example": 1
					},
					"game_match_day": {
						"type": "number",
						"format": "double",
						"description": "Game match day",
						"example": 3
					},
					"game_status": {
						"$ref": "#/components/schemas/GameStatus",
						"description": "Game status",
						"example": "in_progress"
					},
					"game_stage": {
						"$ref": "#/components/schemas/GameStage",
						"description": "Game stage",
						"example": "regular_season"
					},
					"team_id": {
						"type": "string",
						"description": "Team ID",
						"example": "550e8400-e29b-41d4-a716-446655440001"
					},
					"team_name": {
						"type": "string",
						"description": "Team name",
						"example": "Dodgeball Dynamos"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Current game retrieved successfully"
					}
				},
				"required": [
					"game_id",
					"game_season",
					"game_match_day",
					"game_status",
					"game_stage",
					"team_id",
					"team_name",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTeamResponseModel": {
				"description": "Create team response model",
				"properties": {
					"team_id": {
						"type": "string",
						"description": "Team ID",
						"example": "550e8400-e29b-41d4-a716-446655440001"
					},
					"team_name": {
						"type": "string",
						"description": "Team name",
						"example": "Dodgeball Dynamos"
					},
					"game_id": {
						"type": "string",
						"description": "Game ID",
						"example": "550e8400-e29b-41d4-a716-446655440000"
					},
					"game_season": {
						"type": "number",
						"format": "double",
						"description": "Game season",
						"example": 1
					},
					"game_match_day": {
						"type": "number",
						"format": "double",
						"description": "Game match day",
						"example": 1
					},
					"game_status": {
						"$ref": "#/components/schemas/GameStatus",
						"description": "Game status",
						"example": "in_progress"
					},
					"game_stage": {
						"$ref": "#/components/schemas/GameStage",
						"description": "Game stage",
						"example": "draft"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Team created successfully"
					}
				},
				"required": [
					"team_id",
					"team_name",
					"game_id",
					"game_season",
					"game_match_day",
					"game_status",
					"game_stage",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTeamRequest": {
				"description": "Create team request model",
				"properties": {
					"name": {
						"type": "string",
						"description": "Team name",
						"example": "Dodgeball Dynamos"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CancelGameResponseModel": {
				"description": "Cancel game response model",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Game cancelled successfully"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetFacilityInfoResponseModel": {
				"description": "Get facility info response model for tsoa documentation",
				"properties": {
					"training_facility_level": {
						"type": "number",
						"format": "double"
					},
					"scout_level": {
						"type": "number",
						"format": "double"
					},
					"stadium_size": {
						"type": "number",
						"format": "double"
					},
					"training_facility_upgrade_cost": {
						"type": "number",
						"format": "double"
					},
					"scout_upgrade_cost": {
						"type": "number",
						"format": "double"
					},
					"stadium_upgrade_cost": {
						"type": "number",
						"format": "double"
					},
					"can_afford_training_upgrade": {
						"type": "boolean"
					},
					"can_afford_scout_upgrade": {
						"type": "boolean"
					},
					"can_afford_stadium_upgrade": {
						"type": "boolean"
					},
					"budget": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Facility info retrieved successfully"
					}
				},
				"required": [
					"training_facility_level",
					"scout_level",
					"stadium_size",
					"training_facility_upgrade_cost",
					"scout_upgrade_cost",
					"stadium_upgrade_cost",
					"can_afford_training_upgrade",
					"can_afford_scout_upgrade",
					"can_afford_stadium_upgrade",
					"budget",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpgradeFacilityRequestModel": {
				"description": "Upgrade facility request model",
				"properties": {
					"facility_type": {
						"type": "string",
						"enum": [
							"training",
							"scout",
							"stadium"
						],
						"description": "Type of facility to upgrade",
						"example": "training"
					}
				},
				"required": [
					"facility_type"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "dodgeball-manager-api",
		"version": "1.0.0",
		"description": "Dodgeball manager game API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/training/training-info": {
			"get": {
				"operationId": "GetSeasonTrainingInfo",
				"responses": {
					"200": {
						"description": "Season training info retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSeasonTrainingInfoResponseModel"
								}
							}
						}
					}
				},
				"description": "Get training information for the current season",
				"tags": [
					"Training"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/training/train-player": {
			"post": {
				"operationId": "TrainPlayer",
				"responses": {
					"200": {
						"description": "Player trained successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrainPlayerResponseModel"
								}
							}
						}
					}
				},
				"description": "Train a player by improving one of their stats",
				"tags": [
					"Training"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Details of the player to train",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TrainPlayerRequestModel",
								"description": "Details of the player to train"
							}
						}
					}
				}
			}
		},
...
		"/scouting/scouting-info": {
			"get": {
				"operationId": "GetSeasonScoutingInfo",
				"responses": {
					"200": {
						"description": "Season scouting info retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSeasonScoutingInfoResponseModel"
								}
							}
						}
					}
				},
				"description": "Get scouting information for the current season",
				"tags": [
					"Scouting"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/scouting/scouted-players": {
			"get": {
				"operationId": "GetScoutedPlayers",
				"responses": {
					"200": {
						"description": "Scouted players retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetScoutedPlayersResponseModel"
								}
							}
						}
					}
				},
				"description": "Get all scouted players for the current season",
				"tags": [
					"Scouting"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/scouting/scout-players": {
			"post": {
				"operationId": "ScoutPlayers",
				"responses": {
					"200": {
						"description": "Scouted players generated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ScoutPlayersResponseModel"
								}
							}
						}
					}
				},
				"description": "Generate scouted players for the current season",
				"tags": [
					"Scouting"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Details of the scouted players to generate",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ScoutPlayersRequestModel",
								"description": "Details of the scouted players to generate"
							}
						}
					}
				}
			}
		},
		"/scouting/purchase-scouted-player": {
			"post": {
				"operationId": "PurchaseScoutedPlayer",
				"responses": {
					"200": {
						"description": "Scouted player purchased successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseScoutedPlayerResponseModel"
								}
							}
						}
					}
				},
				"description": "Purchase a scouted player for the team",
				"tags": [
					"Scouting"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Details of the scouted player to purchase",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PurchaseScoutedPlayerRequestModel",
								"description": "Details of the scouted player to purchase"
							}
						}
					}
				}
			}
		},
...
		"/games/current": {
			"get": {
				"operationId": "GetCurrentGame",
				"responses": {
					"200": {
						"description": "Current game retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrentGameResponseModel"
								}
							}
						}
					}
				},
				"description": "Get the current game for the authenticated user",
				"summary": "Get current game",
				"tags": [
					"Games"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/games/teams": {
			"post": {
				"operationId": "CreateTeam",
				"responses": {
					"201": {
						"description": "Team created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateTeamResponseModel"
								}
							}
						}
					}
				},
				"description": "Create a new team and game for the authenticated user",
				"summary": "Create new team and game",
				"tags": [
					"Games"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamRequest"
							}
						}
					}
				}
			}
		},
		"/games/cancel": {
			"post": {
				"operationId": "CancelGame",
				"responses": {
					"200": {
						"description": "Game cancelled successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CancelGameResponseModel"
								}
							}
						}
					}
				},
				"description": "Cancel an active game for the authenticated user",
				"summary": "Cancel active game",
				"tags": [
					"Games"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/facilities/facility-info": {
			"get": {
				"operationId": "GetFacilityInfo",
				"responses": {
					"200": {
						"description": "Facility info retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetFacilityInfoResponseModel"
								}
							}
						}
					}
				},
				"description": "Get facility information",
				"tags": [
					"Facilities"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/facilities/upgrade-facility": {
			"post": {
				"operationId": "UpgradeFacility",
				"responses": {
					"200": {
						"description": "Facility upgraded successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpgradeFacilityResponseModel"
								}
							}
						}
					}
				},
				"description": "Upgrade a facility",
				"tags": [
					"Facilities"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Details of the facility to upgrade",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpgradeFacilityRequestModel",
								"description": "Details of the facility to upgrade"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}